Fixed  test cases to be in its own file
