+++ jvm/src/test/kotlin/org/amshove/kluent/tests/assertions/file/ShouldExistShould.kt
@@ -1,8 +1,6 @@
 package org.amshove.kluent.tests.assertions.file
 
-import org.amshove.kluent.Exist
 import org.amshove.kluent.internal.assertFails
-import org.amshove.kluent.should
 import org.amshove.kluent.shouldExist
 import org.junit.Test
 import java.io.File
@@ -18,15 +16,5 @@ class ShouldExistShould {
     @Test
     fun failWhenTestingAFileThatDoesNotExist() {
         assertFails { file.shouldExist() }
-    }
-
-    @Test
-    fun passWhenTestingAFileThatExistsBacktick() {
-        file.useFile { it should Exist }
-    }
-
-    @Test
-    fun failWhenTestingAFileThatDoesNotExistBacktick() {
-        assertFails { file should Exist }
     }
 }
