+++ jvm/src/test/kotlin/org/amshove/kluent/tests/assertions/file/ShouldNotBeAFileShould.kt
@@ -1,7 +1,5 @@
 package org.amshove.kluent.tests.assertions.file
 
-import org.amshove.kluent.aFile
-import org.amshove.kluent.`should not be`
 import org.amshove.kluent.shouldNotBeFile
 import java.io.File
 import kotlin.test.Test
@@ -19,15 +17,5 @@ class ShouldNotBeAFileShould {
     @Test
     fun failWhenTestingAFileThatIsAFile() {
         file.useFile { assertFails { it.shouldNotBeFile() } }
-    }
-
-    @Test
-    fun passWhenTestingAFileThatIsNotAFileBacktick() {
-        dir.useDir { it `should not be` aFile }
-    }
-
-    @Test
-    fun failWhenTestingAFileThatIsAFileBacktick() {
-        file.useFile { assertFails { it `should not be` aFile } }
     }
 }
