+++ jvm/src/test/kotlin/org/amshove/kluent/tests/assertions/file/ShouldNotBeDirShouldBacktick.kt
@@ -1,12 +1,13 @@
 package org.amshove.kluent.tests.assertions.file
 
-import org.amshove.kluent.shouldNotBeDir
+import org.amshove.kluent.Dir
+import org.amshove.kluent.`should not be`
 import org.junit.Before
 import org.junit.Test
 import java.io.File
 import kotlin.test.assertFails
 
-class ShouldNotBeDirShould {
+class ShouldNotBeDirShouldBacktick {
     lateinit var dir: File
     lateinit var file: File
 
@@ -17,12 +18,12 @@ class ShouldNotBeDirShould {
     }
 
     @Test
-    fun passWhenTestingAFile() {
-        dir.useFile { it.shouldNotBeDir() }
+    fun passWhenTestingAFileBacktick() {
+        dir.useFile { it `should not be` Dir }
     }
 
     @Test
-    fun failWhenTestingADir() {
-        file.useDir { assertFails { it.shouldNotBeDir() } }
+    fun failWhenTestingADirBacktick() {
+        file.useDir { assertFails { it `should not be` Dir } }
     }
 }
