+++ jvm/src/test/kotlin/org/amshove/kluent/tests/assertions/file/ShouldNotExistShould.kt
@@ -1,7 +1,5 @@
 package org.amshove.kluent.tests.assertions.file
 
-import org.amshove.kluent.Exist
-import org.amshove.kluent.`should not`
 import org.amshove.kluent.shouldNotExist
 import org.junit.Test
 import java.io.File
@@ -9,7 +7,6 @@ import kotlin.test.assertFails
 
 class ShouldNotExistShould {
     val file = File("test")
-
 
     @Test
     fun passWhenTestingAFileThatDoesNotExist() {
@@ -19,16 +16,5 @@ class ShouldNotExistShould {
     @Test
     fun failWhenTestingAFileThatExists() {
         assertFails { file.useFile { it.shouldNotExist() } }
-    }
-
-
-    @Test
-    fun passWhenTestingAFileThatDoesNotExistBacktick() {
-        file `should not` Exist
-    }
-
-    @Test
-    fun failWhenTestingAFileThatExistsBacktick() {
-        assertFails { file.useFile { it `should not` Exist } }
     }
 }
